+++ src/main/java/site/dogether/auth/service/AuthService.java
@@ -1,0 +1,46 @@
+package site.dogether.auth.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import site.dogether.auth.controller.request.LoginRequest;
+import site.dogether.auth.controller.request.WithdrawRequest;
+import site.dogether.auth.infrastructure.AppleOAuthProvider;
+import site.dogether.auth.infrastructure.JwtHandler;
+import site.dogether.member.domain.Member;
+import site.dogether.member.service.MemberService;
+import site.dogether.member.service.dto.AuthenticatedMember;
+
+@Slf4j
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+@Service
+public class AuthService {
+
+    private final JwtHandler jwtHandler;
+    private final AppleOAuthProvider appleOAuthProvider;
+    private final MemberService memberService;
+
+    @Transactional
+    public AuthenticatedMember login(final LoginRequest request) {
+        final String subject = appleOAuthProvider.getSubjectFromIdToken(request.idToken());
+
+        Member member = new Member(
+                request.idToken(),
+                request.name()
+        );
+        member = memberService.save(member);
+
+        final String token = jwtHandler.createToken(member.getId());
+
+        return new AuthenticatedMember(member.getName(), token);
+    }
+
+    @Transactional
+    public void withdraw(final String token, final WithdrawRequest request) {
+        final Long memberId = jwtHandler.getMemberId(token);
+        memberService.delete(memberId);
+    }
+
+}
